Implement content pack uninstallation (#5058)
* Create helper method in EntityV1 to create a virtual root entity

* Clarify comment about insertion order

* Add EntityFacade#loadNativeEntity() method

This is used to load the native entity for the given descriptor from the
database.

* First working version of content pack uninstall

This hasn't been tested extensively, yet.

* Make sure to never delete the default stream in StreamFacade#delete()

The default stream might be included in a content pack but it should
never be deleted when a content pack gets uninstalled.

* Fix ReferenceConverter when creating a "long" value reference

Jackson actually creates an int value instead of a long when
deserializing a small number so we have to check for long AND int values
when the valueType is LONG.

This also adds some unit tests for the ReferenceConverter.

* Use sort function to sort content pack requirements

Fixes an exception when opening the details of a content pack with more
than one requirement.

+++ graylog2-server/src/main/java/org/graylog2/contentpacks/facades/StreamFacade.java
@@ -33,6 +33,7 @@ import org.graylog2.contentpacks.model.entities.EntityExcerpt;
 import org.graylog2.contentpacks.model.entities.EntityV1;
 import org.graylog2.contentpacks.model.entities.EntityWithConstraints;
 import org.graylog2.contentpacks.model.entities.NativeEntity;
+import org.graylog2.contentpacks.model.entities.NativeEntityDescriptor;
 import org.graylog2.contentpacks.model.entities.StreamEntity;
 import org.graylog2.contentpacks.model.entities.StreamRuleEntity;
 import org.graylog2.contentpacks.model.entities.references.ValueReference;
@@ -213,7 +214,21 @@ public class StreamFacade implements EntityFacade<Stream> {
     }
 
     @Override
+    public Optional<NativeEntity<Stream>> loadNativeEntity(NativeEntityDescriptor nativeEntityDescriptor) {
+        try {
+            final Stream stream = streamService.load(nativeEntityDescriptor.id().id());
+            return Optional.of(NativeEntity.create(nativeEntityDescriptor, stream));
+        } catch (NotFoundException e) {
+            return Optional.empty();
+        }
+    }
+
+    @Override
     public void delete(Stream nativeEntity) {
+        if (nativeEntity.isDefaultStream()) {
+            LOG.debug("The default stream should not be deleted");
+            return;
+        }
         try {
             streamService.destroy(nativeEntity);
         } catch (NotFoundException ignore) {
